account fund works in the Debug.Log. not yet saved in file

+++ Blood And Sand/Assets/Scripts/PlayerScripts/InventoryController.cs
@@ -205,7 +205,7 @@ class Player_Data
 
     //for all items
     //[DataMember]
-	public float item_value;
+	public int item_value;
 
 	//for weapon items	
 	//[DataMember]
